pyproject.toml
[tool.poetry]
name = "classwork"
version = "0.1.0"
description = ""
authors = ["Replit user <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.28.1"
Pillow = "^9.3.0"
Pillow-PIL = "^0.1dev"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
